Responsive y validaciones del formulario de registro. LP
