--- client/src/component/regis/Regist.module.css
@@ -1,135 +1,0 @@
-.registro {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  
-    background-color: #15171e;
-  
-    width: 100%;
-    height: 100vh;
-  }
-  
-  .form {
-    width: 30%;
-    height: 80vh;
-  
-    padding: 3% 0 3% 0;
-  }
-  
-  .logo {
-    background-image: url("../../img/Logo1.png");
-    background-size: 100%;
-    background-repeat: no-repeat;
-    background-position: 50%;
-  
-    border-radius: 10px;
-  
-    width: 70%;
-    height: 22%;
-  
-    margin: 0 auto 10% auto;
-  }
-  
-  .divInputs {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-around;
-    align-items: center;
-    width: 100%;
-    height: 26%;
-  }
-  
-  .input {
-    background-color: rgba(0, 0, 0, 0);
-    color: rgb(255, 255, 255);
-  
-    font-size: 1rem;
-  
-    border: 1px solid rgba(255, 255, 255, 0.36);
-    border-radius: 3px;
-  
-    width: 96%;
-    padding: 8px;
-  
-    transition: 0.3s;
-  }
-  
-  .input::placeholder {
-    color: rgb(156, 156, 156);
-  }
-  
-  .input:focus {
-    color: #fff;
-    outline: 1px solid rgb(20 142 255);
-  }
-  
-  .btn {
-    background-color: #b4b4b4;
-    color: rgb(26, 118, 224);
-  
-    font-size: 1rem;
-  
-    border-radius: 5px;
-    border: 2px solid rgb(95, 160, 235);
-  
-    width: 70%;
-  
-    padding: 2% 0 2% 0;
-    margin: 4% auto 2% auto;
-  
-    transition: 0.1s;
-  }
-  
-  .btn:hover {
-    background-color: #ebebeb;
-    color: rgb(0, 132, 255);
-    outline: 2px solid rgb(26, 118, 224);
-  }
-  
-  h2 {
-    font: 1rem sans-serif;
-    margin-top: 30px;
-    text-align: center;
-  }
-  h2.linea {
-    position: relative;
-    z-index: 1;
-  }
-  h2.linea:before {
-    border-top: 2px solid #fff;
-    content: "";
-    margin: 0 auto;
-    position: absolute;
-    top: 50%;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    width: 95%;
-    z-index: -1;
-  }
-  h2.linea span {
-    color: white;
-    background: #15171e;
-    padding: 0 15px;
-  }
-  
-  .google {
-    background-color: #fff;
-    background-image: url("../../img/google.png");
-    background-size: 42%;
-    background-repeat: no-repeat;
-    background-position: 50%;
-  
-    border-radius: 8px;
-  
-    width: 50%;
-    height: 10%;
-  
-    margin: 12% auto 0 auto;
-  }
-  
-  .pDanger {
-    color: red;
-    font-size: 1rem;
-    margin-top: 0;
-  }
